+++ src/main/java/by/asyncbtd/tl/objects/business/BusinessOpeningHoursInterval.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package by.asyncbtd.tl.objects;
+package by.asyncbtd.tl.objects.business;
 
 public class BusinessOpeningHoursInterval {
